requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.11
alembic==1.12.1
python-multipart==0.0.6
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.1.3
psycopg2-binary==2.9.9